% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept.R
\name{accept_reject}
\alias{accept_reject}
\title{Accept}
\usage{
accept_reject(cur_y, new_y, temp)
}
\arguments{
\item{cur_y}{Current objective}

\item{new_y}{Proposed objective}

\item{temp}{temperature}
}
\value{
Boolean accepting or rejecting proposal
}
\description{
Accept or reject a new solution at a given temperature
}

