% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afford.R
\name{affordable}
\alias{affordable}
\title{Afford}
\usage{
affordable(solution, cost, budget_block, budget, free)
}
\arguments{
\item{solution}{Proposed solution (vector of indices)}

\item{cost}{Solution costs}

\item{budget_block}{index of budget blocks}

\item{budget}{Budget}

\item{free}{Free funds}
}
\value{
Boolean indicating if proposed solution is affordable or not
}
\description{
Check a solution is affordable
}
