% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar.r
\name{covar}
\alias{covar}
\title{Covariance}
\usage{
covar(x, y)
}
\arguments{
\item{x}{A numeric matrix or vector.}

\item{y}{A vector (when \code{x} is a vector) or missing (blank) when 
\code{x} is a matrix.}
}
\value{
The covariance.
}
\description{
Compute covariance efficiently.
}
\details{
This implementation uses the internals for \code{cosine()}.
}
\examples{
library(fastco)
x <- matrix(rnorm(10*3), 10, 3)
pcor(x)

pcor(x[, 1], x[, 2])

}
\seealso{
\code{\link{cosine}}
}

