% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_temp.R
\name{decay}
\alias{decay}
\title{Estimate the temperature decay}
\usage{
decay(starting_temperature, outer_iterations, proportion_greedy)
}
\arguments{
\item{starting_temperature}{The algorithm starting temperature}

\item{outer_iterations}{The number of temperature steps}

\item{proportion_greedy}{The final proportion of the run to be greedy}
}
\value{
The temperature decay
}
\description{
For a given starting temp and final proportion of greedy moves, calculate the
temperature decay such that the temperature is sufficiently low to ensure
greddy search for the requested final proportion of the algorithm
}
