% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_temp.R
\name{find_starting_temp}
\alias{find_starting_temp}
\title{Starting temperature}
\usage{
find_starting_temp(
  target_acceptance,
  current_solution,
  cur_y,
  max_solution,
  min_solution,
  y,
  cost,
  budget_block,
  budget,
  free
)
}
\arguments{
\item{target_acceptance}{Taget acceptance rate}

\item{current_solution}{Current solution index}

\item{cur_y}{Current solution out}

\item{max_solution}{Max solution index}

\item{min_solution}{Min solution index}

\item{y}{Solution out}

\item{cost}{Solution cost}

\item{budget_block}{Budget blocks}

\item{budget}{Budget}

\item{free}{Free funds}
}
\value{
A starting temperature
}
\description{
Increases starting temperature until the acceptence rate over 500 iterations is
greater than the target acceptance rate
}
