% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_dominated.R
\name{frontier}
\alias{frontier}
\title{Find non-dominated solutions: internal}
\usage{
frontier(x)
}
\arguments{
\item{x}{The raw input data (for a single block).}
}
\value{
A set of block non-dominated solutions.
}
\description{
Internal function for finding the non-dominated solutions in a single
block
}

