% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{large_data}
\alias{large_data}
\title{A larger test dataset}
\format{
A data frame with 1560 rows and 4 variables
}
\usage{
large_data
}
\description{
A larger test dataset
}
\keyword{datasets}
