% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{large_data_budget}
\alias{large_data_budget}
\title{Budget for larger test dataset}
\format{
A data frame with 26 rows and 2 variables
}
\usage{
large_data_budget
}
\description{
Budget for larger test dataset
}
\keyword{datasets}
