% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_solution.R
\name{maxs}
\alias{maxs}
\title{Maximum solution}
\usage{
maxs(min_solution, n)
}
\arguments{
\item{min_solution}{Minimum solution}

\item{n}{number of rows in input data}
}
\value{
Index of maximum solutions
}
\description{
Extracts the index of maximum solutions for each admin
}

