% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move_up}
\alias{move_up}
\title{Move up}
\usage{
move_up(solution, max_solution)
}
\arguments{
\item{solution}{Proposed solution (vector of indices)}

\item{max_solution}{Max solution index}
}
\value{
Solution with one up move
}
\description{
Moves the solution for a single admin unit randomly up
}
