% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_dominated.R
\name{non_dominated}
\alias{non_dominated}
\title{Non-dominated}
\usage{
non_dominated(input_data)
}
\arguments{
\item{input_data}{Data: must contain ISO, DIDE_CODE cost and y columns}
}
\value{
inputdata with dominated solutions removed
}
\description{
Remove non dominated solutions from dataset
}

