% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_dominated.R
\name{non_dominated_country}
\alias{non_dominated_country}
\title{Non-dominated country}
\usage{
non_dominated_country(input_data)
}
\arguments{
\item{input_data}{Data: must contain budget_block, solution_block cost and y columns}
}
\value{
inputdata with dominated solutions removed
}
\description{
Remove non dominated solutions from dataset (country-level)
}
