% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.r
\name{pcor}
\alias{pcor}
\title{Pearson Correlation}
\usage{
pcor(x, y)
}
\arguments{
\item{x}{A numeric matrix or vector.}

\item{y}{A vector (when \code{x} is a vector) or missing (blank) when 
\code{x} is a matrix.}
}
\value{
The \eqn{n\times n} matrix of all pair-wise vector cosine
similarities of the columns.
}
\description{
Compute the pearson correlation efficiently.
}
\details{
This implementation uses the internals for \code{cosine()}.
}
\examples{
library(fastcosim)
x <- matrix(rnorm(10*3), 10, 3)
pcor(x)

pcor(x[, 1], x[, 2])

}
\seealso{
\code{\link{cosine}}
}

