% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare data}
\usage{
prepare(input_data_raw)
}
\arguments{
\item{input_data_raw}{Raw input data}
}
\value{
Sorted, non-dominated data
}
\description{
Prepares data for SA
}

