% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_country}
\alias{prepare_country}
\title{Prepare data country}
\usage{
prepare_country(input_data_raw)
}
\arguments{
\item{input_data_raw}{Raw input data}
}
\value{
Sorted, non-dominated data
}
\description{
Prepares data for SA (country-level)
}
