% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Pretty progress bar}
\usage{
progress(outer_iterations, counter)
}
\arguments{
\item{outer_iterations}{The number of temperature steps}

\item{counter}{The current temperature step}
}
\description{
A pretty progress bar to monitor progress.
}
