% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{sa}
\alias{sa}
\title{Simmulated Annealing}
\usage{
sa(input_data, input_budget, outer, inner, frozen = 5000, free = 0)
}
\arguments{
\item{input_data}{Data: must contain ISO, DIDE_CODE cost and y columns}

\item{input_budget}{Data: must contain ISO and budget column}

\item{outer}{Outer (temperature step) iterations}

\item{inner}{Inner interations}

\item{frozen}{Number of attempts with no improvement in final frozen search}

\item{free}{Free funds}
}
\value{
A list with the solution index and trace
}
\description{
SA algorithm
}

