% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{sample1}
\alias{sample1}
\title{Sample x}
\usage{
sample1(x)
}
\arguments{
\item{x}{A vector}
}
\value{
single sample from the x
}
\description{
Samples x, return x if length(x) = 1, else a single sample from x
}
