% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_data_budget}
\alias{small_data_budget}
\title{Budget for small test dataset}
\format{
A data frame with 4 rows and 2 variables
}
\usage{
small_data_budget
}
\description{
Budget for small test dataset
}
\keyword{datasets}
