% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_utils.r
\name{sparsity}
\alias{sparsity}
\title{Sparsity}
\usage{
sparsity(x, proportion = TRUE)
}
\arguments{
\item{x}{The matrix, stored as an ordinary R matrix or as a "simple
triplet matrix" (from the slam package).}

\item{proportion}{Logical; determines if a proportion or a count should be returned.}
}
\value{
The sparsity of the input matrix, of the form a proportion or
a count.
}
\description{
Show the sparsity (as a count or proportion) of a matrix.  For
example, .99 sparsity means 99\% of the values are zero.
Similarly, a sparsity of 0 means the matrix is fully dense.
}
\details{
The implementation is very efficient for dense matrices.  For
sparse triplet matrices, the count is trivial.
}
\examples{
## Completely sparse matrix
x <- matrix(0, 10, 10)
sparsity(x)

## 15\\\% density / 85\\\% sparsity
x[sample(length(x), size=15)] <- 1
sparsity(x)

}

