% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afford.R
\name{tapplysum.fast}
\alias{tapplysum.fast}
\title{Tapply sum fast}
\usage{
tapplysum.fast(x, groups)
}
\arguments{
\item{x}{Values to be summed}

\item{groups}{Groups (numeric)}
}
\value{
Group sums
}
\description{
Optimised group-sum (lifted from ecoreg package)
}
