% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{updown}
\alias{updown}
\title{Propose solution: up and down}
\usage{
updown(current_solution, max_solution, min_solution, cost, ISO, budget, free)
}
\arguments{
\item{current_solution}{Current solution}

\item{free}{Free funds}
}
\value{
A solution
}
\description{
Proposes an affordable new (or the same) solution with a single up move
and a single down move
}

