% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cosine.r
\name{cosine}
\alias{cosine}
\title{Cosine Similarity}
\usage{
cosine(x, y)
}
\arguments{
\item{x}{A numeric matrix or vector.}

\item{y}{A vector (when \code{x} is a vector) or missing (blank) when
\code{x} is a matrix.}
}
\value{
The \eqn{n\times n} matrix of all pair-wise vector cosine
similarities of the columns.  The implementation is dominated
by a symmetric rank-k update via the BLAS function dsyrk(); see
the package \code{README.md} file for a discussion of the
algorithm implementation and complexity.
}
\description{
Compute the cosine similarity matrix efficiently.  The function
syntax and behavior is largely modeled after that of the
\code{cosine()} function from the \code{lsa} package, although
with a very different implementation.
}
\details{
The function computes the cosine similarity between all column
vectors of the numeric input matrix
}
\examples{
library(fastcosim)
x <- matrix(rnorm(10*3), 10, 3)
cosine(x)

cosine(x[, 1], x[, 2])
}

