% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{fastcosim-package}
\alias{fastcosim-package}
\title{fastcosim}
\description{
A micro-package for computing cosine similarity.  This is
commonly needed in, for example, natural language processing,
where the cosine similarity coefficients of all columns of a
term-document or document-term matrix is needed.
}
\section{Implementation Details}{

The only exported function of this package, \code{cosine()},
takes two forms.  If supplied with two vectors, then the cosine
similarity of the two vectors will be computed.  On the other
hand, if supplied a single matrix, then all pair-wise vector
cosine similarities of the columns are computed.  The storage
options for a "matrix" are ordinary R matrices and "simple
triplet matrix" from the slam package.
}
\author{
Drew Schmidt and Wei-Chen Chen
}
\keyword{package}

