% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_solution.R
\name{mins}
\alias{mins}
\title{Minimum solution}
\usage{
mins(DIDE_CODE)
}
\arguments{
\item{DIDE_CODE}{vector of DIDE_CODEs}
}
\value{
Index of minimum solutions
}
\description{
Extracts the index of minimum solutions for each admin
}
