% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.r
\name{pcor}
\alias{pcor}
\title{Pearson Correlation}
\usage{
pcor(x, y)
}
\arguments{
\item{x}{A numeric matrix or vector.}

\item{y}{A vector (when \code{x} is a vector) or missing (blank) when 
\code{x} is a matrix.}
}
\value{
The pearson correlation matrix.
}
\description{
An optimized, efficient implemntation for computing the pearson
correlation.
}
\details{
See \code{?coop} for implementation details.
}
\examples{
x <- matrix(rnorm(10*3), 10, 3)

coop::pcor(x)
coop::pcor(x[, 1], x[, 2])

}
\author{
Drew Schmidt
}
\seealso{
\code{\link{cosine}}
}

